+++ Sources/ZcashLightClientKit/Transaction/TransactionEncoder.swift
@@ -73,7 +73,7 @@ protocol TransactionEncoder {
     func createProposedTransactions(
         proposal: Proposal,
         spendingKey: UnifiedSpendingKey
-    ) async throws -> [ZcashTransaction.Overview]
+    ) async throws -> [CreatedTransaction]
 
     /// Creates a transaction proposal to fulfill a [ZIP-321](https://zips.z.cash/zip-0321), throwing an exception whenever things are missing. When the provided wallet implementation
     /// doesn't throw an exception, we wrap the issue into a descriptive exception ourselves (rather than using